improved docs
