+++ ai4water/models/tensorflow/__init__.py
@@ -13,6 +13,13 @@ if tf is not None:
     from .private_layers import TransformerBlocks
     from .nbeats_keras import NBeats
     from .tft_layer import TemporalFusionTransformer
-    from ._models import TabTransformer
-    from ._models import FTTransformer
+    from ._functions import MLP
+    from ._functions import LSTM
+    from ._functions import CNN
+    from ._functions import CNNLSTM
+    from ._functions import LSTMAutoEncoder
+    from ._functions import TCN
+    from ._functions import TFT
+    from ._functions import TabTransformer
+    from ._functions import FTTransformer
